Moved files to messaging utils folder and updated import calls
Relates #42
