+++ app/(dashboard)/search/[results]/page.tsx
@@ -3,7 +3,7 @@ import BackButton from '@/components/buttons/BackButton';
 import { getItems } from '@/utils/supabase/getItems';
 import filterItems from '@/utils/supabase/filterItems';
 import searchByName from '@/utils/supabase/searchByName';
-import { PartialItem } from '@/utils/supabase/types';
+import { PartialItem } from '@/types/types';
 
 type ParamsType = {
   query: string;
