+++ components/form/ProfileEdit.tsx
@@ -1,6 +1,6 @@
 'use client';
 import editRow from '@/utils/supabase/editRow';
-import { PartialItem, editProfile } from '@/utils/supabase/types';
+import { PartialItem, editProfile } from '@/types/types';
 import { useEffect, useState } from 'react';
 import { useForm } from 'react-hook-form';
 import ButtonPill from '../buttons/ButtonPill';
