+++ components/messaging/CurrentConversation.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { MessageType } from '@/utils/messaging/messagingTypes';
+import { MessageType } from '@/types/messagingTypes';
 import MessageCard from './MessageCard';
 import MessageForm from './MessageForm';
 import { useContext, useEffect, useState } from 'react';
