+++ utils/supabase/insertRow.ts
@@ -1,5 +1,5 @@
 import newClient from '../../config/supabaseclient';
-import { PartialItem, PartialProfile } from './types';
+import { PartialItem, PartialProfile } from '../../types/types';
 export default async function insertRow(
   table: string,
   InsertValues: PartialItem | PartialProfile
